chore: renamed errors package to huberrors
